+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Changed
+- Now requires Imager as a dependency
+
 ## 2.0.2 - 2018-10-22
 ### Fixed
 - Fixed subfolders support in console command